requirements.txt
Django>=5.0.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
psycopg2-binary>=2.9.9
redis>=5.0.1
celery>=5.3.6
django-redis>=5.4.0
Pillow>=10.2.0
geopy>=2.4.1
python-decouple>=3.8
gunicorn>=21.2.0
django-cors-headers>=4.3.1
drf-spectacular>=0.27.1
requests>=2.31.0
